DocumentViewer and ThumbNailList components
This component runs off of reactJS and is responsible for selecting and viewing a list of documents selected for upload, or from a server somewhere that sends a list of files to view: either as a list of thumbnails, or as full documents. Additional operations can be made on the documents such as renaming, removing or adding (all operations should have user-defined callback methods to act on the data selected on or by this component).